requirements.txt
SQLAlchemy==1.3.16
grpcio==1.28.1
grpcio-tools==1.28.1
protobuf==3.11.3
six==1.14.0
psycopg2==2.8.5